+++ Project Lumina/Packages/com.arongranberg.astar/Core/AstarMath.cs
@@ -1762,6 +1762,78 @@ namespace Pathfinding {
 		}
 
 		/// <summary>
+		/// Closest point on the triangle abc to the point p as barycentric coordinates.
+		///
+		/// See: 'Real Time Collision Detection' by Christer Ericson, chapter 5.1, page 141
+		/// </summary>
+		public static float3 ClosestPointOnTriangleBarycentric (float3 a, float3 b, float3 c, float3 p) {
+			// Check if p is in vertex region outside A
+			var ab = b - a;
+			var ac = c - a;
+			var ap = p - a;
+
+			var d1 = math.dot(ab, ap);
+			var d2 = math.dot(ac, ap);
+
+			// Barycentric coordinates (1,0,0)
+			if (d1 <= 0 && d2 <= 0) {
+				return new float3(1, 0, 0);
+			}
+
+			// Check if p is in vertex region outside B
+			var bp = p - b;
+			var d3 = math.dot(ab, bp);
+			var d4 = math.dot(ac, bp);
+
+			// Barycentric coordinates (0,1,0)
+			if (d3 >= 0 && d4 <= d3) {
+				return new float3(0, 1, 0);
+			}
+
+			// Check if p is in edge region outside AB, if so return a projection of p onto AB
+			var vc = d1 * d4 - d3 * d2;
+			if (d1 >= 0 && d3 <= 0 && vc <= 0) {
+				// Barycentric coordinates (1-v, v, 0)
+				var v = d1 / (d1 - d3);
+				return new float3(1-v, v, 0);
+			}
+
+			// Check if p is in vertex region outside C
+			var cp = p - c;
+			var d5 = math.dot(ab, cp);
+			var d6 = math.dot(ac, cp);
+
+			// Barycentric coordinates (0,0,1)
+			if (d6 >= 0 && d5 <= d6) {
+				return new float3(0, 0, 1);
+			}
+
+			// Check if p is in edge region of AC, if so return a projection of p onto AC
+			var vb = d5 * d2 - d1 * d6;
+			if (d2 >= 0 && d6 <= 0 && vb <= 0) {
+				// Barycentric coordinates (1-v, 0, v)
+				var v = d2 / (d2 - d6);
+				return new float3(1 - v, 0, v);
+			}
+
+			// Check if p is in edge region of BC, if so return projection of p onto BC
+			var va = d3 * d6 - d5 * d4;
+			if ((d4 - d3) >= 0 && (d5 - d6) >= 0 && va <= 0) {
+				var v = (d4 - d3) / ((d4 - d3) + (d5 - d6));
+				return new float3(0, 1 - v, v);
+			} else {
+				// P is inside the face region. Compute the point using its barycentric coordinates (u, v, w)
+				var denom = 1f / (va + vb + vc);
+				var v = vb * denom;
+				var w = vc * denom;
+				return new float3(1 - v - w, v, w);
+
+				// This is equal to: u*a + v*b + w*c, u = va*denom = 1 - v - w;
+				// return a + ab * v + ac * w;
+			}
+		}
+
+		/// <summary>
 		/// Closest point on a triangle when one axis is scaled.
 		///
 		/// Project the triangle onto the plane defined by the projection axis.